MODBULKOPS-560 Fixed coverage
