+++ src/test/java/org/folio/bulkops/util/MarcCsvHelperTest.java
@@ -4,7 +4,6 @@ import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.bulkops.domain.dto.OperationStatusType.APPLY_CHANGES;
 import static org.folio.bulkops.domain.dto.OperationStatusType.REVIEW_CHANGES;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 import lombok.SneakyThrows;
@@ -18,9 +17,7 @@ import org.folio.bulkops.domain.dto.InstanceNoteType;
 import org.folio.bulkops.domain.dto.OperationStatusType;
 import org.folio.bulkops.domain.entity.BulkOperation;
 import org.folio.bulkops.service.InstanceReferenceService;
-import org.folio.bulkops.service.Marc21ReferenceProvider;
 import org.folio.bulkops.service.RuleService;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
@@ -29,13 +26,11 @@ import org.marc4j.marc.impl.LeaderImpl;
 import org.marc4j.marc.impl.RecordImpl;
 import org.marc4j.marc.impl.SubfieldImpl;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 import java.io.FileInputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Set;
 import java.util.UUID;
 
 class MarcCsvHelperTest extends BaseTest {
@@ -226,43 +221,5 @@ class MarcCsvHelperTest extends BaseTest {
 
     var expectedInstanceNotes = "";
     assertThat(new String(res)).contains(expectedInstanceNotes);
-  }
-
-  @Nested
-  @SpringBootTest
-  class MarcCsvHelperOptionNotFoundTest {
-
-    @MockitoBean
-    private Marc21ReferenceProvider marc21ReferenceProvider;
-
-    @Test
-    @SneakyThrows
-    void shouldEnrichCsvWithEmptyValueBecauseOfOptionNotFound() {
-      var fileName = "file.csv";
-      var operationId = UUID.randomUUID();
-      var operation = BulkOperation.builder()
-              .id(operationId)
-              .status(REVIEW_CHANGES)
-              .entityType(EntityType.INSTANCE_MARC)
-              .linkToModifiedRecordsMarcCsvFile(fileName)
-              .build();
-
-      var content = new FileInputStream("src/test/resources/files/instance.csv").readAllBytes();
-
-      when(mappingRulesClient.getMarcBibMappingRules())
-              .thenReturn(Files.readString(Path.of("src/test/resources/files/mappingRulesResponse.json")));
-      when(remoteFileSystemClient.get(fileName))
-              .thenReturn(new FileInputStream("src/test/resources/files/marc_csv_empty_notes.csv"));
-      when(ruleService.getMarcRules(operationId))
-              .thenReturn(new BulkOperationMarcRuleCollection()
-                      .bulkOperationMarcRules(singletonList(new BulkOperationMarcRule().tag("500"))));
-      when(marc21ReferenceProvider.getChangedOptionsSetForCsv(any()))
-              .thenReturn(Set.of("Not notes"));
-
-      var res = marcCsvHelper.enrichCsvWithMarcChanges(content, operation);
-
-      var expectedInstanceNotes = "";
-      assertThat(new String(res)).contains(expectedInstanceNotes);
-    }
   }
 }
